fix column typing and remove unneessary UI
